+++ tests/test_datasets/test_analysis.py
@@ -12,9 +12,9 @@ import pandas
 from pykeen.datasets import Nations
 from pykeen.datasets.analysis import (
     SUBSET_LABELS, calculate_relation_functionality, entity_count_dataframe, entity_relation_co_occurrence_dataframe,
-    relation_cardinalities_types, relation_cardinality_classification, relation_classification, relation_count_dataframe,
+    relation_cardinality_classification, relation_classification, relation_count_dataframe,
 )
-from pykeen.triples.analysis import _get_skyline
+from pykeen.triples.analysis import _get_skyline, relation_cardinalities_types
 
 
 def _old_skyline(xs):
